% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{carros}
\alias{carros}
\title{Testes de estrada de carros}
\format{
Data frame com 234 linhas e 11 variáveis:
\describe{
\item{montadora}{nome da montadora}
\item{modelo}{nome do modelo}
\item{cilindrada}{cilindrada do motor, em litros}
\item{ano}{ano de fabricação}
\item{cilindros}{número de cilindros}
\item{transmicao}{tipo de transmição}
\item{tracao}{tipo de tracao}
\item{urb}{consumo urbano, em milhas por galão}
\item{rod}{consumo rodoviário, em milhas por galão}
\item{combustível}{tipo de combustível}
\item{categoria}{categoria do carro}
}
}
\source{
R package ggplot2.
}
\description{
Este conjunto de dados contém um subconjunto dos dados referente à economia de combustível que a Agência de Proteção do Meio Ambiente (APMA) dos EUA disponibiliza em http://fueleconomy.gov. O conjunto de dados contém apenas modelos que tiveram um novo lançamento a cada ano entre 1999 e 2008 - isso foi usado como um proxy para a popularidade do carro.
}
\references{
Henderson and Velleman (1981), Building multiple regression models interactively. Biometrics, 37, 391–411.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
