% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, percentual = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector or an R object but not a factor coercible to numeric by as.double(x).}

\item{percentual}{logical; if TRUE (default), the coefficient of variation is returned in \%.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the coefficient of variation
}
\description{
This function computes the coefficient of variation of the values in x. If na.rm is TRUE then missing values are removed before computation proceeds.
}
