% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_missing_packages}
\alias{install_missing_packages}
\title{Installation of missing packages}
\usage{
install_missing_packages(pkg)
}
\arguments{
\item{pkg}{a vector containing the names of the packages to be installed.}
}
\value{
No return value, called for side effects
}
\description{
Installation of missing packages
}
