% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.installed}
\alias{is.installed}
\title{Checking if a package is installed}
\usage{
is.installed(pkg)
}
\arguments{
\item{pkg}{a vector containing the names of the packages to be checked.}
}
\value{
a data.frame with the information of the installed packages
}
\description{
Checking if a package is installed
}
