% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start_clean_session}
\alias{start_clean_session}
\title{Cleaning global environment and R search path}
\usage{
start_clean_session()
}
\value{
No return value, called for side effects
}
\description{
This function cleans up the global environment and unload all packages loaded in a previous R session.
}
