% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{resistencia}
\alias{resistencia}
\title{Resistência à compressão.}
\format{
Data.frame com 80 linhas e uma variável:
\describe{
\item{resistência}{resistência à compressão (em psi) de 80 corpos de prova de liga de alumínio-lítio.}
}
}
\description{
Exemplo 6.4 do livro \insertCite{2015Montgomery}{useR}. Dados referentes à resistência à compressão (em psi) de 80 corpos de prova de liga de alumínio-lítio.
}
\references{
\insertRef{2015Montgomery}{useR}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
